FIX coordinate display when the length of numbers is greater than 1
Previously numbers greater than length 1 would not be accounted for and every subsequent row would be incorrectly pushed to the right. Now in the coordinate header numbers are displayed vertically (to keep all tiles as a uniform 1x1 square). The row coordinates (preceding a row of tiles) are now right justified and include appropriate whitespace on the left to account for the largest length number.