Add check_any_win method and associated public winner variable Add InvalidInputError and BoardFullError exceptions Removed private parity variable
The board class will now automatically check if a win has been attained after a tile has been set and will update a public winner variable accordingly.

Additionally added InvalidInputError and BoardFullError exceptions for ease of use in the game driver.

Removed the parity variable as it was never relevant since the board will always be a square and it doesnt affect the methods for win checking.