Add GameDriver
The GameDriver file runs the game in its entirety automatically. It prompts users for the number of players (and their associated character) and the size of the board in order to start playing. it handles errors that may occur when playing as well as keeping track of player timers during gameplay.
Once the game is concluded, the endscreen will be printed and the user will be prompted whether or not they wish to start a new game.